Adding ESLint support to Webpack
