+++ .eslintrc
@@ -1,0 +1,22 @@
+{
+  "parserOptions": {
+    "ecmaVersion": 6,
+    "sourceType": "module",
+    "ecmaFeatures": {
+      "jsx": true,
+      "experimentalObjectRestSpread": true
+    }
+  },
+  "rules": {
+//    "semi": 1,
+//    "no-var": 1,
+//    "prefer-const": 1,
+//    "max-len": ["error", { "ignoreRegExpLiterals": true,
+//      "code": 180,
+//      "ignoreStrings": true,
+//      "ignoreTemplateLiterals": true,
+//      "ignoreComments": true}],
+//    "no-lonely-if": 1,
+//    "object-property-newline": 1
+  }
+}
